+++ packages/fluentui-rc/src/BaseInputTemplate/BaseInputTemplate.tsx
@@ -60,7 +60,7 @@ export default function BaseInputTemplate<
     (e: MouseEvent) => {
       e.preventDefault();
       e.stopPropagation();
-      onChange(options.emptyValue ?? '');
+      onChange(options.emptyValue);
     },
     [onChange, options.emptyValue],
   );