+++ packages/shadcn/src/BaseInputTemplate/BaseInputTemplate.tsx
@@ -52,7 +52,7 @@ export default function BaseInputTemplate<
     (e: MouseEvent) => {
       e.preventDefault();
       e.stopPropagation();
-      onChange(options.emptyValue ?? '');
+      onChange(options.emptyValue);
     },
     [onChange, options.emptyValue],
   );